# Extend Task.FromResult default task optimization to 16 byte unmanaged types