For years, async methods / Task.FromResult has cached a `Task<T>` for `default(T)`, however it was only used when the result value was null.  Earlier in this release we extended that optimization to also use the default task when the value was 1, 2, 4, or 8 bytes (and not a reference type).  This extends that further to also handle types that are 16 bytes, so as to include types like Decimal, Guid, Int128, and DateTimeOffset.

```C#
[Params(42, 0)]
public int Value { get; set; }

[Benchmark]
public async Task<Int128> ReturnValue()
{
    await Task.CompletedTask;
    return Value;
}
```

|      Method |         Toolchain | Value |      Mean | Ratio | Allocated | Alloc Ratio |
|------------ |------------------ |------ |----------:|------:|----------:|------------:|
| ReturnValue | \main\corerun.exe |     0 | 18.347 ns |  1.00 |      80 B |        1.00 |
| ReturnValue |   \pr\corerun.exe |     0 |  8.659 ns |  0.47 |         - |        0.00 |
|             |                   |       |           |       |           |             |
| ReturnValue | \main\corerun.exe |    42 | 17.742 ns |  1.00 |      80 B |        1.00 |
| ReturnValue |   \pr\corerun.exe |    42 | 17.744 ns |  1.00 |      80 B |        1.00 |